+++ src/main/java/aston/polina/config/Config.java
@@ -1,13 +1,86 @@
 package aston.polina.config;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.core.env.Environment;
+import org.springframework.jdbc.datasource.DriverManagerDataSource;
+import org.springframework.orm.hibernate5.HibernateTransactionManager;
+import org.springframework.orm.hibernate5.LocalSessionFactoryBean;
+import org.springframework.transaction.annotation.EnableTransactionManagement;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+import javax.sql.DataSource;
+import java.util.Properties;
+
+
 @Configuration
 @EnableWebMvc
+@EnableTransactionManagement
 @ComponentScan(value = "aston.polina")
+@PropertySource(value = "classpath:application.properties")
 public class Config implements WebMvcConfigurer {
+
+    @Value("${spring.datasource.driver-class-name}")
+    private String DB_DRIVER;
+    @Value("${spring.datasource.url}")
+    private String DB_URL;
+    @Value("${spring.datasource.username}")
+    private String DB_USERNAME;
+    @Value("${spring.datasource.password}")
+    private String DB_PASSWORD;
+
+    private final Environment env;
+
+    @Autowired
+    public Config(Environment env) {
+        this.env = env;
+    }
+
+    @Bean
+    public LocalSessionFactoryBean sessionFactory() {
+        LocalSessionFactoryBean sessionFactory = new LocalSessionFactoryBean();
+        sessionFactory.setDataSource(dataSource());
+        sessionFactory.setPackagesToScan("aston.polina");
+        sessionFactory.setHibernateProperties(hibernateProperties());
+
+        return sessionFactory;
+    }
+
+    public DataSource dataSource() {
+        DriverManagerDataSource dataSource = new DriverManagerDataSource();
+        dataSource.setDriverClassName(DB_DRIVER);
+        dataSource.setUrl(DB_URL);
+        dataSource.setUsername(DB_USERNAME);
+        dataSource.setPassword(DB_PASSWORD);
+
+        return dataSource;
+    }
+
+    @Bean
+    public HibernateTransactionManager hibernateTransactionManager() {
+        HibernateTransactionManager transactionManager
+                = new HibernateTransactionManager();
+        transactionManager.setSessionFactory(sessionFactory().getObject());
+        return transactionManager;
+    }
+
+    private Properties hibernateProperties() {
+        Properties hibernateProperties = new Properties();
+        hibernateProperties.put("spring.jpa.show-sql", env.getProperty("spring.jpa.show-sql"));
+        hibernateProperties.put("spring.jpa.hibernate.ddl-auto", env.getProperty("spring.jpa.hibernate.ddl-auto"));
+        hibernateProperties.put("spring.jpa.hibernate.connection.CharSet",
+                env.getProperty("spring.jpa.hibernate.connection.CharSet"));
+        hibernateProperties.put("spring.jpa.properties.hibernate.dialect",
+                env.getProperty("spring.jpa.properties.hibernate.dialect"));
+        hibernateProperties.put("spring.jpa.properties.hibernate.current_session_context_class",
+                env.getProperty("spring.jpa.properties.hibernate.current_session_context_class"));
+
+        return hibernateProperties;
+    }
 
 }
